turn events into a markdown page
