+++ docs/contributing/events.md
@@ -1,0 +1,10 @@
+---
+title: Gatsby Community Events
+---
+
+Interested in connecting with the Gatsby community in person? Take a look at the list below to see community-organized Gatsby events.
+
+Want to see your event featured here? [Learn more about submitting your event for Gatsby support](/contributing/organize-a-gatsby-event/) (Support can include free swag, \$ for food, and more!)
+
+<Events />
+<EmailCaptureForm signupMessage="Want to keep up with the latest tips & tricks? Subscribe to our newsletter!" />
