+++ www/src/gatsby-plugin-theme-ui/components.js
@@ -12,6 +12,7 @@ import CodeBlock from "../components/code-block"
 import MdxLink from "../components/mdx-link"
 import Breakout from "../components/breakout"
 import VisuallyHidden from "../components/visually-hidden"
+import Events from "../components/events/events"
 
 export default {
   GuideList,
@@ -25,5 +26,6 @@ export default {
   a: MdxLink,
   Breakout,
   VisuallyHidden,
+  Events,
   pre: ({ children }) => <CodeBlock>{children}</CodeBlock>,
 }
